# order-service/src/main/java/com/minicommerce/orders/web/dto/OrderResponse.java
package com.minicommerce.orders.web.dto;

import java.math.BigDecimal;
import java.time.OffsetDateTime;
import java.util.List;
import java.util.UUID;

public record OrderResponse(
        UUID id,
        UUID customerId,
        String status,
        String currency,
        BigDecimal total,
        OffsetDateTime createdAt,
        OffsetDateTime updatedAt,
        List<OrderItemResponse> items
) { }