# order-service/src/main/java/com/minicommerce/orders/web/error/ApiError.java
package com.minicommerce.orders.web.error;

import java.time.OffsetDateTime;

public record ApiError(
        OffsetDateTime timestamp,
        int status,
        String error,
        String message,
        String path
) { }